Create `FEATURE_PORTABLE_HELPERS` (#119414)
Rename USE_PORTABLE_HELPERS to FEATURE_PORTABLE_HELPERS in native AOT.
Add place holder directory for portable implementations under runtime/.
Move empty WASM stubs to portable locations and enable FEATURE_PORTABLE_HELPERS on WASM.
