+++ src/coreclr/clr.featuredefines.props
@@ -4,6 +4,7 @@
         <FeaturePerfTracing>true</FeaturePerfTracing>
         <FeatureEHFunclets>true</FeatureEHFunclets>
         <FeaturePortableEntryPoints>false</FeaturePortableEntryPoints>
+        <FeaturePortableHelpers>false</FeaturePortableHelpers>
         <ProfilingSupportedBuild>true</ProfilingSupportedBuild>
     </PropertyGroup>
 
@@ -13,6 +14,7 @@
 
     <PropertyGroup Condition="'$(TargetsWasm)' == 'true'">
         <FeaturePortableEntryPoints>true</FeaturePortableEntryPoints>
+        <FeaturePortableHelpers>true</FeaturePortableHelpers>
     </PropertyGroup>
 
     <PropertyGroup Condition="'$(TargetsUnix)' == 'true'">
@@ -49,6 +51,7 @@
         <DefineConstants Condition="'$(FeatureEHFunclets)' == 'true'">$(DefineConstants);FEATURE_EH_FUNCLETS</DefineConstants>
         <DefineConstants Condition="'$(FeatureInterpreter)' == 'true'">$(DefineConstants);FEATURE_INTERPRETER</DefineConstants>
         <DefineConstants Condition="'$(FeaturePortableEntryPoints)' == 'true'">$(DefineConstants);FEATURE_PORTABLE_ENTRYPOINTS</DefineConstants>
+        <DefineConstants Condition="'$(FeaturePortableHelpers)' == 'true'">$(DefineConstants);FEATURE_PORTABLE_HELPERS</DefineConstants>
 
         <DefineConstants Condition="'$(ProfilingSupportedBuild)' == 'true'">$(DefineConstants);PROFILING_SUPPORTED</DefineConstants>
     </PropertyGroup>
