+++ src/coreclr/clrdefinitions.cmake
@@ -121,6 +121,9 @@ add_compile_definitions($<${FEATURE_INTERPRETER}:FEATURE_INTERPRETER>)
 if (FEATURE_PORTABLE_ENTRYPOINTS)
   add_compile_definitions(FEATURE_PORTABLE_ENTRYPOINTS)
 endif()
+if (FEATURE_PORTABLE_HELPERS)
+  add_compile_definitions(FEATURE_PORTABLE_HELPERS)
+endif()
 
 if (CLR_CMAKE_TARGET_WIN32)
     add_definitions(-DFEATURE_ISYM_READER)
