+++ src/coreclr/clrfeatures.cmake
@@ -45,6 +45,7 @@ if(NOT DEFINED FEATURE_INTERPRETER)
   elseif(CLR_CMAKE_TARGET_ARCH_WASM)
     set(FEATURE_INTERPRETER 1)
     set(FEATURE_PORTABLE_ENTRYPOINTS 1)
+    set(FEATURE_PORTABLE_HELPERS 1)
   else()
     if(CLR_CMAKE_TARGET_ARCH_AMD64 OR CLR_CMAKE_TARGET_ARCH_ARM64)
       set(FEATURE_INTERPRETER $<IF:$<CONFIG:Debug,Checked>,1,0>)
