+++ src/coreclr/nativeaot/Runtime/startup.cpp
@@ -122,13 +122,13 @@ static bool InitDLL(HANDLE hPalInstance)
         return false;
 
     // Note: The global exception handler uses RuntimeInstance
-#if !defined(USE_PORTABLE_HELPERS)
+#if !defined(FEATURE_PORTABLE_HELPERS)
 #ifdef HOST_WINDOWS
     AddVectoredExceptionHandler(1, RhpVectoredExceptionHandler);
 #else
     PalSetHardwareExceptionHandler(RhpHardwareExceptionHandler);
 #endif
-#endif // !USE_PORTABLE_HELPERS
+#endif // !FEATURE_PORTABLE_HELPERS
 
 #ifdef STRESS_LOG
     uint32_t dwTotalStressLogSize = (uint32_t)g_pRhConfig->GetTotalStressLogSize();
@@ -158,7 +158,7 @@ static bool InitDLL(HANDLE hPalInstance)
     EventPipe_FinishInitialize();
 #endif
 
-#ifndef USE_PORTABLE_HELPERS
+#ifndef FEATURE_PORTABLE_HELPERS
     if (!DetectCPUFeatures())
         return false;
 #endif
@@ -171,7 +171,7 @@ static bool InitDLL(HANDLE hPalInstance)
     return true;
 }
 
-#ifndef USE_PORTABLE_HELPERS
+#ifndef FEATURE_PORTABLE_HELPERS
 
 bool DetectCPUFeatures()
 {
@@ -206,7 +206,7 @@ bool DetectCPUFeatures()
 
     return true;
 }
-#endif // !USE_PORTABLE_HELPERS
+#endif // !FEATURE_PORTABLE_HELPERS
 
 #ifdef TARGET_UNIX
 inline
