+++ src/coreclr/nativeaot/Runtime/windows/PalMinWin.cpp
@@ -195,7 +195,7 @@ uint64_t PalGetCurrentOSThreadId()
     return GetCurrentThreadId();
 }
 
-#if !defined(USE_PORTABLE_HELPERS) && !defined(FEATURE_RX_THUNKS)
+#if !defined(FEATURE_PORTABLE_HELPERS) && !defined(FEATURE_RX_THUNKS)
 UInt32_BOOL PalAllocateThunksFromTemplate(_In_ HANDLE hTemplateModule, uint32_t templateRva, size_t templateSize, _Outptr_result_bytebuffer_(templateSize) void** newThunksOut)
 {
 #ifdef XBOX_ONE
@@ -235,7 +235,7 @@ UInt32_BOOL PalFreeThunksFromTemplate(_In_ void *pBaseAddress, size_t templateSi
     return UnmapViewOfFile(pBaseAddress);
 #endif
 }
-#endif // !USE_PORTABLE_HELPERS && !FEATURE_RX_THUNKS
+#endif // !FEATURE_PORTABLE_HELPERS && !FEATURE_RX_THUNKS
 
 UInt32_BOOL PalMarkThunksAsValidCallTargets(
     void *virtualAddress,
